Skip building binaries in  when there are no tests
