+++ benchmarks/random-access-bench/Cargo.toml
@@ -29,5 +29,9 @@ vortex-bench = { workspace = true }
 lance = ["dep:lance-bench"]
 unstable_encodings = ["vortex-bench/unstable_encodings"]
 
+[[bin]]
+name = "random-access-bench"
+test = false
+
 [lints]
 workspace = true
